Move worker into actor folder.
