+++ crates/worker/src/actor/native_worker.rs
